backend/requirements.txt
asgiref==3.8.1
click==8.1.8
colorama==0.4.6
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2024.2
